Added raw and primitive serializers
