+++ src/Microsoft.Extensions.OData.Migration/Formatters/Serialization/ODataMigrationRawValueSerializer.cs
@@ -1,0 +1,32 @@
+﻿// ------------------------------------------------------------------------------
+// <copyright company="Microsoft Corporation">
+//     Copyright © Microsoft Corporation. All rights reserved.
+// </copyright>
+// ------------------------------------------------------------------------------
+
+namespace Microsoft.Extensions.OData.Migration.Formatters.Serialization
+{
+    using Microsoft.AspNet.OData.Formatter.Serialization;
+    using Microsoft.OData;
+    using System;
+    using System.Collections.Generic;
+    using System.Text;
+    public class ODataMigrationRawValueSerializer : ODataRawValueSerializer
+    {
+        public ODataMigrationRawValueSerializer() : base()
+        {
+        }
+
+        public override void WriteObject(object graph, Type type, ODataMessageWriter messageWriter, ODataSerializerContext writeContext)
+        {
+            if (Type.GetTypeCode(type) == TypeCode.UInt64)
+            {
+                base.WriteObject(graph.ToString(), typeof(string), messageWriter, writeContext);
+            }
+            else
+            {
+                base.WriteObject(graph, type, messageWriter, writeContext);
+            }
+        }
+    }
+}
