+++ src/Microsoft.Extensions.OData.Migration/Microsoft.Extensions.OData.Migration.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.8.45</Version>
+    <Version>1.8.300</Version>
   </PropertyGroup>
 
   <Import Project="..\Build.props" />
